requirements.txt
mcp
python-dotenv
requests
streamlit
openai
